refactor: remove unused type definitions and interfaces, move the extension manager related types and interfaces to the extension directory to be modular
