+++ src/routes/settings/+page.svelte
@@ -4,7 +4,7 @@
     import { getAvailableModifiers, getAvailableKeys, updateShortcut } from '../../utils/shortcutManager';
     import { goto } from '$app/navigation';
     import { settingsService, settings as settingsStore } from '../../services/settingsService';
-    import extensionManager, { extensionUninstallInProgress } from '../../services/extensionManager';
+    import extensionManager, { extensionUninstallInProgress } from '../../services/extension/extensionManager';
     import { get } from 'svelte/store';
   import type { AppSettings } from '../../types';
   import { logService } from '../../services/logService';
