+++ src/services/extension/extensionManager.ts
@@ -1,6 +1,6 @@
 import { writable, get } from "svelte/store";
-import { searchQuery } from "../stores/search";
-import { settingsService } from "./settingsService";
+import { searchQuery } from "../../stores/search";
+import { settingsService } from "../settingsService";
 import { exists, readDir, remove } from "@tauri-apps/plugin-fs";
 import { join, resourceDir, appDataDir } from "@tauri-apps/api/path";
 import { invoke } from "@tauri-apps/api/core";
@@ -13,12 +13,12 @@ import type {
 } from "asyar-api";
 import { discoverExtensions } from "./extensionDiscovery";
 import { ExtensionBridge } from "asyar-api";
-import { logService } from "./logService";
-import { NotificationService } from "./notificationService";
-import { ClipboardHistoryService } from "./clipboardHistoryService";
-import { actionService } from "./actionService";
+import { logService } from "../logService";
+import { NotificationService } from "../notificationService";
+import { ClipboardHistoryService } from "../clipboardHistoryService";
+import { actionService } from "../actionService";
 import { commandService } from "./commandService";
-import { performanceService } from "./performanceService";
+import { performanceService } from "../performanceService";
 
 // Import components
 import {
@@ -29,11 +29,11 @@ import {
   ShortcutRecorder,
   SplitView,
   ConfirmDialog,
-} from "../components";
+} from "../../components";
 // Removed 'Command' import as it's not used directly here after removing old indexing
 // import type { Command } from "./search/types/Command";
-import type { SearchableItem } from "./search/types/SearchableItem";
-import { searchService } from "./search/SearchService";
+import type { SearchableItem } from "../search/types/SearchableItem";
+import { searchService } from "../search/SearchService";
 
 // Stores for extension state
 export const extensionUninstallInProgress = writable<string | null>(null);
@@ -267,7 +267,7 @@ class ExtensionManager implements IExtensionManager {
 
       const extensionPairs = await Promise.all(
         extensionIds.map((id) =>
-          this.loadExtensionWithManifest(`../extensions/${id}`)
+          this.loadExtensionWithManifest(`../../extensions/${id}`)
         )
       );
 
@@ -442,7 +442,7 @@ class ExtensionManager implements IExtensionManager {
       for (const id of extensionIds) {
         try {
           // Directly read manifest without importing the extension module
-          const manifestPath = `../extensions/${id}/manifest.json`;
+          const manifestPath = `../../extensions/${id}/manifest.json`;
           const manifest = await import(/* @vite-ignore */ manifestPath);
 
           if (manifest) {
@@ -672,9 +672,9 @@ class ExtensionManager implements IExtensionManager {
         // Adjust this path based on your actual dev setup relative to `src-tauri`
         // Assuming extensions are at the root level for simplicity here
         basePath = await resourceDir(); // Or use a fixed relative path if more reliable
-        // This might need adjustment: `../extensions` relative to where JS code runs?
+        // This might need adjustment: `../../extensions` relative to where JS code runs?
         logService.warn(
-          "Using development path for extensions. Ensure '../extensions' is correct relative to runtime."
+          "Using development path for extensions. Ensure '../../extensions' is correct relative to runtime."
         );
         return await join(basePath, "..", "extensions"); // Example adjustment
       } else {
