+++ src/services/index.ts
@@ -1,8 +1,8 @@
 // Export all services
 export { logService } from "./logService";
-export { default as extensionManager } from "./extensionManager";
+export { default as extensionManager } from "./extension/extensionManager";
 export { settingsService } from "./settingsService";
 export { ClipboardHistoryService as clipboardService } from "./clipboardHistoryService";
 export { actionService } from "./actionService";
-export { default as commandService } from "./commandService";
+export { default as commandService } from "./extension/commandService";
 // Add other services as needed
