--- src/types/ResultCategoryType.ts
@@ -1,6 +1,0 @@
-export type ResultCategory =
-  | "application"
-  | "command"
-  | "calculation"
-  | "snippet"
-  | "clipboard";
