+++ src/types/index.ts
@@ -1,5 +1,1 @@
-export type { ActionResult } from "./ActionResultType";
 export type { AppSettings } from "./AppSettingsType";
-export type { CategoryResults } from "./CategoryResultsType";
-export type { ResultCategory } from "./ResultCategoryType";
-export type { SearchResults, SearchResultItem } from "./SearchResultsType";
